Add parameterized Terraform action to Jenkinsfile
Refactor Jenkins pipeline to use a parameter for Terraform actions and remove redundant stages.

+++ Jenkinsfile
@@ -1,6 +1,14 @@
 pipeline {
     agent any
 
+    parameters {
+        choice(
+            name: 'ACTION',
+            choices: ['plan', 'apply', 'destroy'],
+            description: 'Choose Terraform action to perform'
+        )
+    }
+
     environment {
         AWS_DEFAULT_REGION = "us-east-1"
     }
@@ -9,14 +17,6 @@ pipeline {
         stage('Checkout') {
             steps {
                 checkout scm
-            }
-        }
-
-        stage('Terraform Version') {
-            steps {
-                ansiColor('xterm') {
-                    sh 'terraform version'
-                }
             }
         }
 
@@ -30,60 +30,21 @@ pipeline {
             }
         }
 
-        stage('Terraform Validate') {
+        stage('Terraform Action') {
             steps {
                 withCredentials([[$class: 'AmazonWebServicesCredentialsBinding', credentialsId: 'aws-creds']]) {
                     ansiColor('xterm') {
-                        sh 'terraform validate'
-                    }
-                }
-            }
-        }
-
-        stage('Terraform Plan') {
-            steps {
-                withCredentials([[$class: 'AmazonWebServicesCredentialsBinding', credentialsId: 'aws-creds']]) {
-                    ansiColor('xterm') {
-                        sh 'terraform plan -out=tfplan'
-                    }
-                }
-            }
-        }
-
-        stage('Approval') {
-            steps {
-                input 'Apply Terraform?'
-            }
-        }
-
-        stage('Terraform Apply') {
-            steps {
-                withCredentials([[$class: 'AmazonWebServicesCredentialsBinding', credentialsId: 'aws-creds']]) {
-                    ansiColor('xterm') {
-                        sh 'terraform apply -auto-approve tfplan'
-                    }
-                }
-            }
-        }
-
-        stage('Destroy Approval') {
-            steps {
-                input 'Destroy Terraform resources?'
-            }
-        }
-
-        stage('Terraform Destroy') {
-            when {
-                beforeAgent true
-                expression {
-                    // Only run if user approved the Destroy stage
-                    true
-                }
-            }
-            steps {
-                withCredentials([[$class: 'AmazonWebServicesCredentialsBinding', credentialsId: 'aws-creds']]) {
-                    ansiColor('xterm') {
-                        sh 'terraform destroy -auto-approve'
+                        script {
+                            if (params.ACTION == 'plan') {
+                                sh 'terraform plan -out=tfplan'
+                            } else if (params.ACTION == 'apply') {
+                                input message: 'Are you sure you want to APPLY changes?'
+                                sh 'terraform apply -auto-approve tfplan'
+                            } else if (params.ACTION == 'destroy') {
+                                input message: 'Are you sure you want to DESTROY all resources?'
+                                sh 'terraform destroy -auto-approve'
+                            }
+                        }
                     }
                 }
             }
